Add system from SMIRNOFF class method
